Moving WeatherListLocationSearch to the WeatherListViewModel. - Small refactor to the WeatherListActivity to use the databinding items instead of re-declaring them.
